// src/app/enums/permissoes.ts
export enum Permissoes {
  GESTAO_FORMULARIO = 'GESTAO_FORMULARIO',
  MODELO_CONSULTA = 'MODELO_CONSULTA',
  ELABORACAO_ETP = 'ELABORACAO_ETP',
  ETP_CONSULTA = 'ETP_CONSULTA',
  ASSUNTO = 'ADMINISTRACAO_ETP',
  SITUACAO = 'ADMINISTRACAO_ETP',
  TIPO_CONTRATACAO = 'ADMINISTRACAO_ETP',
  ETAPAS_ETP = 'ADMINISTRACAO_ETP',
  NUMERACAO_ETP = 'ADMINISTRACAO_ETP',
  ROTULOS = 'ADMINISTRACAO_ETP',
  ETP_TIPO_PERMISSAO = 'ADMINISTRACAO_ETP',
  TIPO_DELEGACAO = 'ADMINISTRACAO_ETP',
  ETP_BLOQUEADOS = 'ADMINISTRACAO_ETP',
  FORMULARIOS_BLOQUEADOS = 'ADMINISTRACAO_ETP',
  PRIORIDADE = 'ADMINISTRACAO_ETP',
}